+++ cmd/cfv/cfv.go
@@ -736,12 +736,13 @@ func formatDefaults(formatName string) formatter.Options {
 		}
 	case "yaml":
 		return formatter.Options{
-			IndentStyle:  formatter.IndentSpaces,
-			IndentWidth:  2,
-			FinalNewline: true,
-			LineEnding:   formatter.LineEndingLF,
-			QuoteStyle:   formatter.QuoteDouble,
-			SortKeys:     false,
+			IndentStyle:     formatter.IndentSpaces,
+			IndentWidth:     2,
+			FinalNewline:    true,
+			LineEnding:      formatter.LineEndingLF,
+			QuoteStyle:      formatter.QuoteDouble,
+			SortKeys:        false,
+			IndentSequences: formatter.SequenceIndentEnabled,
 		}
 	case "jsonc":
 		return formatter.Options{
@@ -816,6 +817,13 @@ func applyFormatOptions(opts *formatter.Options, cfg *configfile.FormatOptions) 
 			opts.TrailingCommas = formatter.TrailingCommasNone
 		default:
 			opts.TrailingCommas = formatter.TrailingCommasPreserve
+		}
+	}
+	if cfg.IndentSequences != nil {
+		if *cfg.IndentSequences {
+			opts.IndentSequences = formatter.SequenceIndentEnabled
+		} else {
+			opts.IndentSequences = formatter.SequenceIndentDisabled
 		}
 	}
 }