+++ pkg/formatter/yamlfmt/yaml.go
@@ -36,10 +36,11 @@ var _ formatter.Formatter = Formatter{}
 // DefaultOptions returns the default formatting options for YAML.
 func DefaultOptions() formatter.Options {
 	return formatter.Options{
-		IndentStyle:  formatter.IndentSpaces,
-		IndentWidth:  2,
-		FinalNewline: true,
-		QuoteStyle:   formatter.QuoteDouble,
+		IndentStyle:     formatter.IndentSpaces,
+		IndentWidth:     2,
+		FinalNewline:    true,
+		QuoteStyle:      formatter.QuoteDouble,
+		IndentSequences: formatter.SequenceIndentEnabled,
 	}
 }
 
@@ -113,6 +114,9 @@ func resolveOptions(opts formatter.Options) formatter.Options {
 	}
 	if opts.IndentWidth == 0 {
 		opts.IndentWidth = defaults.IndentWidth
+	}
+	if opts.IndentSequences == formatter.SequenceIndentDefault {
+		opts.IndentSequences = defaults.IndentSequences
 	}
 	return opts
 }